requirements.txt
requests
selenium
pillow
tldextract
dnspython
ttkbootstrap
